+++ src/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import './index.css';
 import { App} from './App';
+import { AppMobile } from './mobile-app/AppMobile';
 import { app } from "./components/app";
 import user from "./models/user";
 import inspiration from "./models/inspiration";
@@ -9,5 +10,10 @@ import inspiration from "./models/inspiration";
 app.model(user);
 app.model(inspiration);
 
-app.router(() => <App />);
+function isMobileDevice() {
+    return (typeof window.orientation !== "undefined") || (navigator.userAgent.indexOf('IEMobile') !== -1);
+};
+
+true ?  app.router(() => <AppMobile />) :  app.router(() => <App />);
+
 app.start('#root');
