+++ src/mobile-app/components/Navigation/AppHeaderStyle.css
