+++ src/mobile-app/components/Pages/Profile/Avatar.js
