+++ src/mobile-app/components/Pages/Profile/profileHelper.js
